Update README.md
GUIESPFLASHING is a powerful tool designed to facilitate the flashing and management of firmware on ESP-based devices. With the latest updates, the software now includes a set of enhanced features to make the flashing process more intuitive and user-friendly. These updates include the ability to view the web console of the ESP device, visual indicators for each step in the firmware flashing process, and an integrated Chromium-based web browser for accessing the device's web view.